refactor: move follow button to a component
